+++ src/sass/_index.scss
@@ -1,4 +1,4 @@
-@forward './base-rules.scss';
+@forward './root-layout.scss';
 @forward './blueprint-overrides.scss';
 @forward './pages';
 @forward './abstracts';
