--- src/sass/base-rules.scss
@@ -1,32 +1,0 @@
-@import './abstracts';
-.warning-banner {
-  color: red;
-  text-align: center;
-}
-
-// Main body
-.content-wrapper {
-  display: flex;
-  justify-content: center;
-  padding-bottom: 4rem; // Footer height + some padding
-
-  @media screen and (max-width: $mediaMobileWidth) {
-    #content {
-      padding: 0 0.5em 0 0.5em;
-    }
-  }
-}
-
-footer {
-  height: 2.5rem;
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  text-align: center;
-}
-
-.flex-container {
-  display: flex;
-  gap: 1rem;
-  flex-wrap: wrap;
-}
